+++ .eslintrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  root: true,
+  extends: ['cutting/eslint-config/node']
+};
