+++ .gitignore
@@ -3,3 +3,7 @@
 !/.idea/runConfigurations/*.xml
 
 /node_modules/**
+
+dist
+
+*.tsbuildinfo
