+++ package.json
@@ -1,17 +1,19 @@
 {
   "name": "ajv-json-loader",
-  "version": "0.1.4",
+  "version": "0.2.0",
   "description": "Webpack loader that compiles JSON strings into Ajv validation functions using ajv-pack",
   "engines": {
-    "node": ">=6.4.0"
+    "node": ">=14.0.0"
   },
   "files": [
-    "index.js"
+    "dist"
   ],
   "scripts": {
-    "test": "mocha"
+    "prepack": "tsc",
+    "pretest": "yarn prepack",
+    "test": "TS_NODE_PROJECT=\"test.tsconfig.json\" mocha -r ts-node/register test/**/*.test.ts"
   },
-  "main": "index.js",
+  "main": "dist/index.js",
   "repository": {
     "type": "git",
     "url": "git://git@github.com/downace/ajv-json-loader.git"
@@ -32,15 +34,26 @@
   },
   "homepage": "https://github.com/downace/ajv-json-loader#readme",
   "dependencies": {
-    "ajv": "~6.5.0",
-    "ajv-pack": "~0.3.1",
-    "loader-utils": "~1.1.0"
+    "ajv": "^8.8.2",
+    "assert-ts": "^0.3.4",
+    "loader-utils": "^2.0.2"
   },
   "devDependencies": {
-    "chai": "~4.1.2",
-    "chai-string": "~1.4.0",
-    "memory-fs": "~0.4.1",
-    "mocha": "~5.2.0",
-    "webpack": "~3.8.1"
+    "@cutting/eslint-config": "^4.28.1",
+    "@cutting/tsconfig": "^4.26.1",
+    "@cutting/useful-types": "^4.25.5",
+    "@types/chai": "^4.3.0",
+    "@types/chai-string": "^1.4.2",
+    "@types/loader-utils": "^2.0.3",
+    "@types/memory-fs": "^0.3.3",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^17.0.0",
+    "chai": "^4.3.4",
+    "chai-string": "^1.5.0",
+    "memory-fs": "^0.5.0",
+    "mocha": "^9.1.3",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.5.4",
+    "webpack": "^4.46.0"
   }
 }
